# docs/cmdb-api-spec.yaml
openapi: 3.0.1
info:
  title: CMDB API
  description: This is Chelas Movies Database API
  contact:
    email: cmdbWebSite@isel.ipl.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - description: Localhost server
    url: http://localhost:1904/

tags:
- name: Users
- name: Movies
- name: Group
- name: Movies in a Group

paths:
  /api/users:
    post:
      tags:
      - Users
      summary: Add a user to the system
      description: Adds a user to the system
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        description: user to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUser'
        required: true
      responses:
        201:
          description: user created
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/CreatedUser'
        400:
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorized'
                example: "Invalid User: already exists"        
      x-codegen-request-body-name: user
  /api/movies:
    get:
      tags:
      - Movies
      summary: Get the most popular movies
      description: Search for the most popular movies
      operationId: getMovies
      security:
        - bearerAuth: []
      parameters:
      - name: limit
        in: query
        description: maximum number of movies to return per page, has a default value of 10
        schema:
          minimum: 1
          type: integer
      - name: page
        in: query
        description: page number, has a default value of 1
        schema:
          minimum: 1
          type: integer          
      responses:
        200:
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
        400:
          description: bad input parameter
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidRequest'
                example: "Invalid Argument: limit"
        404:
          description: argument not found
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/RequestNotFound'
                example: "Argument not found: page"
  /api/movies/search/{moviesName}:
    get:
      tags:
      - Movies
      summary: Search movies by name
      description: Search movies by name
      operationId: searchMoviesByName
      security:
        - bearerAuth: []
      parameters:
      - name: moviesName
        in: path
        description: movies to search
        required: true
        schema:
          type: string
      - name: limit
        in: query
        description: maximum number of movies to return per page, has a default value of 10
        schema:
          minimum: 1
          type: integer
      - name: page
        in: query
        description: page number, has a default value of 1
        schema:
          minimum: 1
          type: integer          
      responses:
        200:
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
        400:
          description: bad input parameter
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidRequest'
                example: "Invalid Argument: limit"
        404:
          description: argument not found
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/RequestNotFound'
                example: "Argument not found: page"
  /api/movies/find/{movieId}:
    get:
      tags:
      - Movies
      summary: Get movie details by id
      description: Given an id retrieve movie details
      operationId: getMovieDetails
      security:
        - bearerAuth: []
      parameters:
      - name: movieId
        in: path
        description: movie to retrieve details from
        required: true
        schema:
          type: string
      responses:
        200:
          description: movie details result
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MovieDetails'
  /api/groups:
    get:
      tags:
      - Group
      summary: List all groups
      description: List all groups
      operationId: getAllGroups
      security:
        - bearerAuth: []
      parameters:
      - name: limit
        in: query
        description: maximum number of groups to return per page, has a default value of 6
        schema:
          minimum: 1
          type: integer
      - name: page
        in: query
        description: page number, has a default value of 1
        schema:
          minimum: 1
          type: integer          
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        401:
          description: Invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorized'
        400:
          description: bad input parameter
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidRequest'
                example: "Invalid Argument: limit"
        404:
          description: argument not found
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/RequestNotFound'
                example: "Argument not found: page"
    post:
      tags:
      - Group
      summary: Add a new group
      description: adds a new group
      operationId: addGroup
      security:
        - bearerAuth: []
      requestBody:
        description: Group to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGroup'
        required: true
      responses:
        201:
          description: Group created
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/NewGroupCreated'
        400:
          description: Group missing a valid name and description
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidRequest'
                example: "Invalid Argument: group is missing a valid name and description"
        401:
          description: Invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorized'
  /api/groups/{groupId}:
    get:
      tags:
      - Group
      summary: List the group details
      description: Get all movies in a group
      operationId: getGroupDetails
      security:
        - bearerAuth: []
      parameters:
      - name: groupId
        in: path
        description: Group of movies
        required: true
        schema:
          type: string
      - name: limit
        in: query
        description: maximum number of movies to return per page, has a default value of 9
        schema:
          minimum: 1
          type: integer
      - name: page
        in: query
        description: page number, has a default value of 1
        schema:
          minimum: 1
          type: integer          
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        404:
          description: Group not found
          content: 
            application/json:
              schema: 
                $ref: '#/components/schemas/RequestNotFound'
                example: "Argument not found: group"
        401:
          description: Invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorized'        
    put:
      tags:
      - Group
      summary: Edit a group
      description: Edit a Group
      operationId: editGroup
      security:
        - bearerAuth: []
      parameters:
      - name: groupId
        in: path
        description: Group to be edited
        required: true
        schema:
          type: string
      requestBody:
        description: Group to be edited
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewGroup'
        required: true
      responses:
        200:
          description: Group updated
          content:
            application/json:
              schema:
                required:
                - message
                type: object
                properties:
                  message:
                    type: string
                    example: "Group updated with success"
        400:
          description: Group missing a valid name and description
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidRequest'
                example: "Invalid Argument: group is missing a valid name and description"
        404:
          description: Group not found
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/RequestNotFound'
                example: "Argument not found: group"
        401:
          description: Invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorized'        
    delete:
      tags:
      - Group
      summary: Remove a group
      description: Remove a group by ID
      operationId: removeGroup
      security:
        - bearerAuth: []
      parameters:
      - name: groupId
        in: path
        description: Group of movies
        required: true
        schema:
          type: string
      responses:
        200:
          description: Group removed with success
          content: 
            application/json:
              schema: 
                $ref: '#/components/schemas/ValidObjectRemoved'
        404:
          description: Group not found
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/RequestNotFound'
                example: "Argument not found: group"
        401:
          description: Invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorized'
  /api/groups/{groupId}/movies/{movieId}:
    put:
      tags:
      - Movies in a Group
      summary: Add a new movie to a group
      description: add a new movie to the group given
      operationId: addMovie
      security:
        - bearerAuth: []
      parameters:
      - name: groupId
        in: path
        description: Group of movies
        required: true
        schema:
          type: string
      - name: movieId
        in: path
        description: A single Movie
        required: true
        schema:
          type: string
      responses:
        201:
          description: movie added
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/NewMovieCreated'
        400:
          description: Movie is already added in the group
          content: 
            application/json:
              schema: 
                $ref: '#/components/schemas/InvalidRequest'
                example: "Invalid Argument: movie already exists"
        404:
          description: Movie or Group not found
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/RequestNotFound'
                example: "Argument not found: movie or group"
        401:
          description: Invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorized'        
    delete:
      tags:
      - Movies in a Group
      summary: Remove a movie from a group
      description: Remove a movie in the group given
      operationId: removeMovie
      security:
        - bearerAuth: []
      parameters:
      - name: groupId
        in: path
        description: Group of movies
        required: true
        schema:
          type: string
      - name: movieId
        in: path
        description: A single Movie
        required: true
        schema:
          type: string
      responses:
        200:
          description: Movie removed with success
          content: 
            application/json:
              schema: 
                $ref: '#/components/schemas/ValidObjectRemoved'
        404:
          description: Movie or Group not found
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/RequestNotFound'
                example: "Argument not found: movie"          
        401:
          description: Invalid authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorized'        
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Group:
      required:
      - name
      - description
      - id
      - movies
      - userId
      type: object
      properties:
        name:
          type: string
          example: Favourite Movies
        description:
          type: string
          example: Group description
        id:
          type: integer
          minimum: 0
          example: 4
        movies:
          type: array
          items: 
            $ref: '#/components/schemas/Movie'
        userId:
          type: integer
    NewGroup:
      required:
      - name
      - description
      type: object
      properties:
        name:
          type: string
          example: Favourite Movies
        description:
          type: string
          example: Group description
    NewGroupCreated:
      required:
      - message
      - group
      type: object
      properties:
        message:
          type: string
          example: Group created
        group:
          $ref: '#/components/schemas/Group'
    Movie:
      required:
      - id
      - name
      - runtimeMins
      type: object
      properties:
        id:
          type: string
          example: "tt02345678"
        name:
          type: string
          example: Jurassic Park
        runtimeMins:
          type: integer
          minimum: 0
          example: 180
    MovieDetails:
      required:
      - id
      - title
      - description
      - image_url
      - runtimeMins
      - director
      - actors_names
      type: object
      properties:
        id:
          type: string
          example: "tt02345678"
        title:
          type: string
          example: "The Dark Knight"
        description:
          type: string
          example: "When the menace known as the Joker wreaks havoc.."
        image_url:
          type: string 
          example: "https://m.media-amazon.com/images/M/MV5BMTMxNTMwODM0NF5BMl5BanBnXkFtZTcwODAyMTk2Mw@@._V1_Ratio0.6762_AL_.jpg"
        runtimeMins:
          type: integer
          minimum: 0
          example: 152      
        director:
          type: string
          example: "Christopher Nolan"
        actors_names:
          type: string
          example: "Christian Bale, Heath Ledger, Aaron Eckhart"
    NewMovieCreated:
      required:
      - status
      - movie
      type: object
      properties:
        status:
          type: string
          example: Movie with name "Jurassic Park" created with success
        movie:
          $ref: '#/components/schemas/Movie'
    ValidObjectRemoved:
      type: object
      properties:
        message:
          type: string
          example: "Removed with success"
    NewUser:
      required:
      - username
      - email
      - password
      - passConfirm
      type: object
      properties:
        username:
          type: string
          example: "guest"
        email:
          type: string
          example: "guest@gmail.com"
        password:
          type: string
          example: "abcd"  
        passConfirm:
          type: string
          example: "abcd"
    CreatedUser:
      type: object
      properties:
          id: 
            type: string 
          username:
            type: string
          email:
            type: string
            example: "guest@gmail.com"
          password:
            type: string
          token:
            type: string
            format: uuid
    NotAuthorized:
      type: object
      properties:
        error:
          type: string
          example: "Invalid authentication token"
    InvalidRequest:
      type: object
      properties:
        error:
          type: string
          example: "Missing required parameter"
    RequestNotFound:
      type: object
      properties:
        error:
          type: string
          example: "Request not found"